+++ src/board/rules.rs
@@ -10,37 +10,21 @@ use crate::types::Piece;
 pub struct Rules;
 
 impl Rules {
-    // pub fn x(b: &Board) {
-    //     let mut capture_mask = Bitboard::all();
-    //     let mut push_mask    = Bitboard::all();
-    //     let checkers = b.checkers_of(b.color_us());
+ 
 
-    //     if checkers.popcount() == 1 {
-    //         // if ony one checker, we can evade check by capturing it
-    //         capture_mask = checkers;
-    //         // If the piece giving check is a slider, we can evade check by blocking it
-    //         if board.line_pieces().intersects(checkers) {
-    //             push_mask = opponent_slider_rays_to_square(king_square, board);
-    //         } else {
-    //             // if the piece is not a slider, we can only evade check by capturing
-    //             push_mask = Bitboard::EMPTY;
-    //         }
-    //     }
+    // pub fn pseudo_legals(b: &Board, moves: &mut MoveList) {
+    //     Rules::pawn_captures_incl_promo(b, moves);
+    //     Rules::pawn_promos(b, moves);
+    //     Rules::pawn_push(b, moves);
+    //     // Rules::pawn_capture_promos(b, moves);
+    //     Rules::non_pawn(Piece::Knight, b, moves);
+    //     Rules::non_pawn(Piece::Bishop, b, moves);
+    //     Rules::non_pawn(Piece::Rook, b, moves);
+    //     Rules::non_pawn(Piece::Queen, b, moves);
+    //     // Rules::non_pawn(Piece::King,b,  moves);
+    //     Rules::king_legal(b, moves);
+    //     Rules::castles(b, moves);
     // }
-
-    pub fn pseudo_legals(b: &Board, moves: &mut MoveList) {
-        Rules::pawn_captures_incl_promo(b, moves);
-        Rules::pawn_promos(b, moves);
-        Rules::pawn_push(b, moves);
-        // Rules::pawn_capture_promos(b, moves);
-        Rules::non_pawn(Piece::Knight, b, moves);
-        Rules::non_pawn(Piece::Bishop, b, moves);
-        Rules::non_pawn(Piece::Rook, b, moves);
-        Rules::non_pawn(Piece::Queen, b, moves);
-        // Rules::non_pawn(Piece::King,b,  moves);
-        Rules::king_legal(b, moves);
-        Rules::castles(b, moves);
-    }
 
     pub fn king_legal(b: &Board, moves: &mut MoveList) {
         let us = b.us();
@@ -55,8 +39,6 @@ impl Rules {
         let king_sq = (b.kings() & us).square();
         let king_att = attack_gen.king_attacks(king_sq);
         let king_danger = BoardCalcs::threats_to(b, b.color_us(), occ - our_kings);
-        // let attacks =
-        //     attack_gen.non_pawn_attacks(b.color_us(), Piece::King, us, them, king_sq) & !us - king_danger;
         let attacks = king_att & !us - king_danger;
         moves.extend(attacks.squares().map(|to| {
             if to.is_in(them) {
@@ -67,38 +49,22 @@ impl Rules {
         }))
     }
 
-    // // e/p pawn_captures
-    // let ep = b.en_passant();
-    // if !ep.is_empty() {
-    //     let ep = ep.square();
-    //     if ep.is_in(pawn_captures_e) {
-    //         let from = ep.shift(c.pawn_capture_east().opposite());
-    //         let capture_square = ep.shift(c.opposite().forward());
-    //         let m =Move::new_ep_capture(from, ep, capture_square);
-    //         moves.push(m);
-    //     }
-    //     if ep.is_in(pawn_captures_w) {
-    //         let from = ep.shift(c.pawn_capture_west().opposite());
-    //         let capture_square = ep.shift(c.opposite().forward());
-    //         let m =Move::new_ep_capture(from, ep, capture_square);
-    //         moves.push(m);
-    //     }
 
     pub fn legals_for(b: &Board, moves: &mut MoveList) {
         let us = b.us();
         let them = b.them();
         let our_kings = b.kings() & us;
-        if our_kings.is_empty() {
-            return Self::pseudo_legals(b, moves);
-        }
+        // if our_kings.is_empty() {
+        //     return Self::pseudo_legals(b, moves);
+        // }
         let checkers = b.checkers_of(b.color_us());
         if checkers.popcount() >= 2 {
             return Self::king_legal(b, moves);
         }
         let the_checker = checkers.square();
-        let king_sq = our_kings.square();
         let gen = BitboardDefault::default();
         if checkers.popcount() == 1 {
+            let king_sq = our_kings.square();
             // Captr checker -> capture-sq & piece not pinned & (checker)
             // Block checker -> to & piece not pinned & (xrays to checker excl)
             let blocking = gen.between(king_sq, the_checker) | checkers; // "| checkers" is for knight checkers
@@ -112,11 +78,17 @@ impl Rules {
             Self::king_legal(b, moves);
             return;
         } else {
+            let pinned = b.pinned();
+            let king_sq = if pinned.is_empty() {
+                Square::null()
+            } else {
+                our_kings.square()
+            };
             for &p in Piece::ALL_BAR_KING.iter() {
                 // not in check
                 for fr in (b.pieces(p) & us).squares() {
                     let attacks = gen.non_pawn_attacks(b.color_us(), p, us, them, fr) & !us;
-                    if !fr.is_in(b.pinned()) {
+                    if !fr.is_in(pinned) {
                         // all non pinned pieces
                         Self::add_moves(attacks, p, fr, b, moves);
                     } else {
@@ -181,13 +153,13 @@ impl Rules {
     #[inline]
     fn add_moves(dests: Bitboard, p: Piece, fr: Square, b: &Board, moves: &mut MoveList) {
         if p != Piece::Pawn {
-            moves.extend(dests.squares().map(|to| {
+            for to in dests.squares() {
                 if to.is_in(b.them()) {
-                    Move::new_capture(p, fr, to, b.piece_at(to.as_bb()))
+                    moves.push(Move::new_capture(p, fr, to, b.piece_at(to.as_bb())))
                 } else {
-                    Move::new_quiet(p, fr, to)
+                    moves.push(Move::new_quiet(p, fr, to))
                 }
-            }))
+            }
         } else {
             if Bitboard::PROMO_RANKS.contains(dests) {
                 Self::add_moves_pawn_promo(dests, fr, b, moves);
@@ -237,14 +209,14 @@ impl Rules {
         }
     }
 
-    pub fn non_pawn(p: Piece, b: &Board, moves: &mut MoveList) {
-        let attack_gen = BitboardDefault::default();
-        let us = b.us();
-        for fr in (b.pieces(p) & us).squares() {
-            let attacks = attack_gen.non_pawn_attacks(b.color_us(), p, b.us(), b.them(), fr) & !us;
-            Self::add_moves(attacks, p, fr, b, moves);
-        }
-    }
+    // pub fn non_pawn(p: Piece, b: &Board, moves: &mut MoveList) {
+    //     let attack_gen = BitboardDefault::default();
+    //     let us = b.us();
+    //     for fr in (b.pieces(p) & us).squares() {
+    //         let attacks = attack_gen.non_pawn_attacks(b.color_us(), p, b.us(), b.them(), fr) & !us;
+    //         Self::add_moves(attacks, p, fr, b, moves);
+    //     }
+    // }
 
     pub fn castles(b: &Board, moves: &mut MoveList) {
         let c = b.color_us();
@@ -294,101 +266,101 @@ impl Rules {
         }
     }
 
-    pub fn pawn_promos(b: &Board, moves: &mut MoveList) {
-        let attack_gen = BitboardDefault::default();
-        let c = b.color_us();
-        let occ = b.occupied();
-        let us = b.us();
-        let pawn_push = attack_gen.pawn_pushes(occ, b.pawns() & us, c);
-        for to in (pawn_push & Bitboard::PROMO_RANKS).squares() {
-            let from = to.shift(c.forward().opposite());
-            // try and pre-sort promos by likely usefulness
-            moves.extend(
-                [Piece::Queen, Piece::Knight, Piece::Rook, Piece::Bishop]
-                    .iter()
-                    .map(|&p| Move::new_promo(from, to, p)),
-            );
-        }
-    }
+    // pub fn pawn_promos(b: &Board, moves: &mut MoveList) {
+    //     let attack_gen = BitboardDefault::default();
+    //     let c = b.color_us();
+    //     let occ = b.occupied();
+    //     let us = b.us();
+    //     let pawn_push = attack_gen.pawn_pushes(occ, b.pawns() & us, c);
+    //     for to in (pawn_push & Bitboard::PROMO_RANKS).squares() {
+    //         let from = to.shift(c.forward().opposite());
+    //         // try and pre-sort promos by likely usefulness
+    //         moves.extend(
+    //             [Piece::Queen, Piece::Knight, Piece::Rook, Piece::Bishop]
+    //                 .iter()
+    //                 .map(|&p| Move::new_promo(from, to, p)),
+    //         );
+    //     }
+    // }
 
-    pub fn pawn_push(b: &Board, moves: &mut MoveList) {
-        // non-promoted single-push pawns
-        let attack_gen = BitboardDefault::default();
-        let c = b.color_us();
-        let occ = b.occupied();
-        let us = b.us();
-        let pawns = b.pawns() & us;
-        let pawn_push = attack_gen.pawn_pushes(occ, pawns, c);
-        for to in (pawn_push & !Bitboard::PROMO_RANKS).squares() {
-            let behind = to.shift(c.forward().opposite());
-            if !behind.is_in(pawns) {
-                // must have been double push
-                let ep = behind;
-                let from = behind.shift(c.forward().opposite());
-                let m = Move::new_double_push(from, to, ep);
-                moves.push(m);
-            } else {
-                let m = Move::new_quiet(Piece::Pawn, behind, to);
-                // FIXME!!!! // { from, to, ep, mover: Piece::Pawn, ..Default::default() };
-                moves.push(m);
-            }
-        }
-    }
+    // pub fn pawn_push(b: &Board, moves: &mut MoveList) {
+    //     // non-promoted single-push pawns
+    //     let attack_gen = BitboardDefault::default();
+    //     let c = b.color_us();
+    //     let occ = b.occupied();
+    //     let us = b.us();
+    //     let pawns = b.pawns() & us;
+    //     let pawn_push = attack_gen.pawn_pushes(occ, pawns, c);
+    //     for to in (pawn_push & !Bitboard::PROMO_RANKS).squares() {
+    //         let behind = to.shift(c.forward().opposite());
+    //         if !behind.is_in(pawns) {
+    //             // must have been double push
+    //             let ep = behind;
+    //             let from = behind.shift(c.forward().opposite());
+    //             let m = Move::new_double_push(from, to, ep);
+    //             moves.push(m);
+    //         } else {
+    //             let m = Move::new_quiet(Piece::Pawn, behind, to);
+    //             // FIXME!!!! // { from, to, ep, mover: Piece::Pawn, ..Default::default() };
+    //             moves.push(m);
+    //         }
+    //     }
+    // }
 
-    pub fn pawn_captures_incl_promo(b: &Board, moves: &mut MoveList) {
-        let attack_gen = BitboardDefault::default();
-        let c = b.color_us();
-        let us = b.us();
-        let them = b.them();
-        let pawns = b.pawns() & us;
+    // pub fn pawn_captures_incl_promo(b: &Board, moves: &mut MoveList) {
+    //     let attack_gen = BitboardDefault::default();
+    //     let c = b.color_us();
+    //     let us = b.us();
+    //     let them = b.them();
+    //     let pawns = b.pawns() & us;
 
-        let (pawn_captures_e, pawn_captures_w) = attack_gen.pawn_attacks(pawns, c);
-        for to in (pawn_captures_e & them).squares() {
-            let from = to.shift(c.pawn_capture_east().opposite());
-            let captured = b.piece_at(to.as_bb());
-            if to.is_in(Bitboard::PROMO_RANKS) {
-                moves.extend(
-                    [Piece::Queen, Piece::Knight, Piece::Rook, Piece::Bishop]
-                        .iter()
-                        .map(|&p| Move::new_promo_capture(from, to, p, captured)),
-                );
-            } else {
-                let m = Move::new_capture(Piece::Pawn, from, to, captured);
-                moves.push(m);
-            }
-        }
-        for to in (pawn_captures_w & them).squares() {
-            let from = to.shift(c.pawn_capture_west().opposite());
-            let captured = b.piece_at(to.as_bb());
-            if to.is_in(Bitboard::PROMO_RANKS) {
-                moves.extend(
-                    [Piece::Queen, Piece::Knight, Piece::Rook, Piece::Bishop]
-                        .iter()
-                        .map(|&p| Move::new_promo_capture(from, to, p, captured)),
-                );
-            } else {
-                let m = Move::new_capture(Piece::Pawn, from, to, captured);
-                moves.push(m);
-            }
-        }
-        // e/p pawn_captures
-        let ep = b.en_passant();
-        if !ep.is_empty() {
-            let ep = ep.square();
-            if ep.is_in(pawn_captures_e) {
-                let from = ep.shift(c.pawn_capture_east().opposite());
-                let capture_square = ep.shift(c.opposite().forward());
-                let m = Move::new_ep_capture(from, ep, capture_square);
-                moves.push(m);
-            }
-            if ep.is_in(pawn_captures_w) {
-                let from = ep.shift(c.pawn_capture_west().opposite());
-                let capture_square = ep.shift(c.opposite().forward());
-                let m = Move::new_ep_capture(from, ep, capture_square);
-                moves.push(m);
-            }
-        }
-    }
+    //     let (pawn_captures_e, pawn_captures_w) = attack_gen.pawn_attacks(pawns, c);
+    //     for to in (pawn_captures_e & them).squares() {
+    //         let from = to.shift(c.pawn_capture_east().opposite());
+    //         let captured = b.piece_at(to.as_bb());
+    //         if to.is_in(Bitboard::PROMO_RANKS) {
+    //             moves.extend(
+    //                 [Piece::Queen, Piece::Knight, Piece::Rook, Piece::Bishop]
+    //                     .iter()
+    //                     .map(|&p| Move::new_promo_capture(from, to, p, captured)),
+    //             );
+    //         } else {
+    //             let m = Move::new_capture(Piece::Pawn, from, to, captured);
+    //             moves.push(m);
+    //         }
+    //     }
+    //     for to in (pawn_captures_w & them).squares() {
+    //         let from = to.shift(c.pawn_capture_west().opposite());
+    //         let captured = b.piece_at(to.as_bb());
+    //         if to.is_in(Bitboard::PROMO_RANKS) {
+    //             moves.extend(
+    //                 [Piece::Queen, Piece::Knight, Piece::Rook, Piece::Bishop]
+    //                     .iter()
+    //                     .map(|&p| Move::new_promo_capture(from, to, p, captured)),
+    //             );
+    //         } else {
+    //             let m = Move::new_capture(Piece::Pawn, from, to, captured);
+    //             moves.push(m);
+    //         }
+    //     }
+    //     // e/p pawn_captures
+    //     let ep = b.en_passant();
+    //     if !ep.is_empty() {
+    //         let ep = ep.square();
+    //         if ep.is_in(pawn_captures_e) {
+    //             let from = ep.shift(c.pawn_capture_east().opposite());
+    //             let capture_square = ep.shift(c.opposite().forward());
+    //             let m = Move::new_ep_capture(from, ep, capture_square);
+    //             moves.push(m);
+    //         }
+    //         if ep.is_in(pawn_captures_w) {
+    //             let from = ep.shift(c.pawn_capture_west().opposite());
+    //             let capture_square = ep.shift(c.opposite().forward());
+    //             let m = Move::new_ep_capture(from, ep, capture_square);
+    //             moves.push(m);
+    //         }
+    //     }
+    // }
 }
 
 #[cfg(test)]